move schema-related utils out of app_manager.util
app_manager.app_schemas.*
