+++ corehq/apps/api/resources/v0_4.py
@@ -23,7 +23,7 @@ from corehq.apps.api.resources.meta import CustomResourceMeta
 from corehq.apps.api.resources.v0_1 import _safe_bool
 from corehq.apps.api.serializers import CommCareCaseSerializer, XFormInstanceSerializer
 from corehq.apps.api.util import get_object_or_not_exist, get_obj
-from corehq.apps.app_manager import util as app_manager_util
+from corehq.apps.app_manager.app_schemas import get_case_properties
 from corehq.apps.app_manager.dbaccessors import get_apps_in_domain
 from corehq.apps.app_manager.models import Application, RemoteApp
 from corehq.apps.cloudcare.api import ElasticCaseQuery
@@ -372,7 +372,7 @@ class ApplicationResource(CouchResourceMixin, HqBaseResource, DomainSpecificReso
 
             dehydrated['case_type'] = module.case_type
 
-            dehydrated['case_properties'] = app_manager_util.get_case_properties(
+            dehydrated['case_properties'] = get_case_properties(
                 app, [module.case_type], defaults=['name']
             )[module.case_type]
