+++ corehq/apps/app_manager/app_schemas/__init__.py
@@ -1,0 +1,3 @@
+from case_properties import ParentCasePropertyBuilder, get_case_properties, \
+    get_all_case_properties, get_usercase_properties, all_case_properties_by_domain,\
+    get_per_type_defaults
