+++ corehq/apps/app_manager/views/view_generic.py
@@ -2,6 +2,8 @@ from django.http import Http404
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.shortcuts import render
+from corehq.apps.app_manager.app_schemas import get_all_case_properties, \
+    get_usercase_properties
 from corehq.apps.app_manager.const import APP_V1
 
 from corehq.apps.app_manager.views.modules import get_module_template, \
@@ -29,9 +31,7 @@ from corehq.apps.hqmedia.views import (
     ProcessAudioFileUploadView,
 )
 from corehq.apps.app_manager.util import (
-    get_all_case_properties,
     get_commcare_versions,
-    get_usercase_properties,
     get_app_manager_template,
 )
 from corehq import toggles
