+++ corehq/apps/data_interfaces/views.py
@@ -9,7 +9,7 @@ from django.http import HttpResponseBadRequest
 
 from corehq import privileges, toggles
 from corehq.apps.accounting.decorators import requires_privilege_with_fallback
-from corehq.apps.app_manager.util import all_case_properties_by_domain
+from corehq.apps.app_manager.app_schemas import all_case_properties_by_domain
 from corehq.apps.casegroups.dbaccessors import get_case_groups_in_domain, \
     get_number_of_case_groups_in_domain
 from corehq.apps.casegroups.models import CommCareCaseGroup
