+++ corehq/apps/reminders/views.py
@@ -7,6 +7,7 @@ from django.utils.decorators import method_decorator
 import pytz
 from django.urls import reverse
 from django.http import HttpResponseRedirect, Http404, HttpResponse
+from corehq.apps.app_manager.app_schemas import get_case_properties
 from corehq.apps.app_manager.dbaccessors import get_apps_in_domain
 from corehq.apps.style.decorators import use_datatables, use_jquery_ui, \
     use_timepicker, use_select2
@@ -18,7 +19,6 @@ from django.utils.translation import ugettext as _, ugettext_noop, ugettext_lazy
 from corehq import privileges
 from corehq.apps.accounting.decorators import requires_privilege_with_fallback
 from corehq.apps.app_manager.models import Form
-from corehq.apps.app_manager.util import get_case_properties
 from corehq.apps.hqwebapp.views import (CRUDPaginatedViewMixin,
     DataTablesAJAXPaginationMixin)
 from corehq import toggles
