+++ corehq/apps/reports/models.py
@@ -20,10 +20,10 @@ from django.utils.translation import ugettext_noop
 from jsonfield import JSONField
 
 from sqlalchemy.util import immutabledict
+from corehq.apps.app_manager.app_schemas import get_case_properties
 
 from corehq.apps.app_manager.dbaccessors import get_app
 from corehq.apps.app_manager.models import Form, RemoteApp
-from corehq.apps.app_manager.util import get_case_properties
 from corehq.apps.cachehq.mixins import (
     CachedCouchDocumentMixin,
     QuickCachedDocumentMixin,
